requirements_simple.txt
# RadiografIA Pro - Versión Simplificada
# Dependencias optimizadas para Windows 10 con TensorFlow 2.10
# TensorFlow 2.10 optimizado para CPU
tensorflow==2.10.1
# Procesamiento de imágenes esencial (compatible con TF 2.10)
opencv-python>=4.5.0,<4.9.0
Pillow>=8.0.0,<11.0.0
# Computación científica básica (compatible con TF 2.10)
numpy>=1.21.0,<1.25.0
scikit-learn>=1.0.0,<1.4.0
# Interfaz gráfica
PyQt5>=5.15.0
# Manipulación de datos básica (compatible con TF 2.10)
pandas>=1.3.0,<2.1.0
# Utilidades esenciales
tqdm>=4.60.0
# Visualización básica (compatible con TF 2.10)
matplotlib>=3.5.0,<3.8.0